# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: TeamVault API
  version: 0.9.1
  description: |
    Zero-Knowledge Passwortmanager REST-API.
    Der Server speichert nur Ciphertext + Envelopes; Ver-/Entschlüsselung
    erfolgt ausschließlich in WebUI, Extension oder CLI (cryptocore).

    API-Key-Scopes (mind. einer): `read` (GET-Allowlist), `vault` (Secret-/Vault-Mutationen),
    `admin` (`/api/admin/*`, plus User-Rollen). Cookie-Sessions ohne Scope-Limit.

    Instanz-Backup: `GET /api/admin/backup`, `POST /api/admin/backup/restore` (platform_admin, Ciphertext only).
servers:
  - url: http://127.0.0.1:8080
    description: Local default
...
security:
  - sessionCookie: []
  - bearerAPIKey: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tv_session
    bearerAPIKey:
      type: http
      scheme: bearer
      bearerFormat: tvk_
      description: |
        API-Key aus Admin-UI (einmalig angezeigt). Gebunden an User+Tenant.
        Scopes: read | vault | admin (mind. einer).
...
paths:
...
  /api/secrets/group-share-gaps:
    get:
      tags: [secrets]
      summary: Group-shared secrets missing envelopes for members (caller must have access)
      description: |
        Used to auto-reseal after new group membership. Optional filters group_id, user_id.
        Response includes the caller's envelope so the client can open the data key (ZK).
      parameters:
        - name: group_id
          in: query
          schema: { type: string }
        - name: user_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: items to reseal
...